requirements.txt
altair==4.2.2
appdirs==1.4.4
attrs==22.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.0
blinker==1.5
cachetools==5.3.0
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.1
decorator==5.1.1
entrypoints==0.4
frozendict==2.3.6
gitdb==4.0.10
GitPython==3.1.31
gunicorn==20.1.0
html5lib==1.1
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
Jinja2==3.1.2
jsonschema==4.17.3
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
multitasking==0.0.11
numpy==1.21.6
packaging==23.0
pandas==1.3.5
patsy==0.5.3
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
plotly==5.14.0
plotly-express==0.4.1
protobuf==3.20.3
pyarrow==11.0.0
pycparser==2.21
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rich==13.3.3
scipy==1.7.3
semver==2.13.0
six==1.16.0
smmap==5.0.0
soupsieve==2.4
statsmodels==0.13.5
streamlit==1.20.0
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
webencodings==0.5.1
wincertstore==0.2
yfinance==0.2.14
zipp==3.15.0
